# Optimize dynamical scatting factor calculation