@sezelt 2022-11-11
Also adding an option to round off the _g_ vector magnitudes for computing the atomic form factors so that the caching does not get messed up by floating point errors, and to allow some nice speedup at the cost of some precision if the user wants to go fast.